Prepare for variadic bind.
