+++ include/boost/phoenix/bind/detail/cpp03/preprocessed/function_ptr.hpp
@@ -5,19 +5,19 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ==============================================================================*/
 
-#if !defined(BOOST_PHOENIX_PREPROCESSED_BIND_FUNCTION_OBJECT_HPP)
-#define BOOST_PHOENIX_PREPROCESSED_BIND_FUNCTION_OBJECT_HPP
+#if !defined(BOOST_PHOENIX_PREPROCESSED_BIND_DETAIL_FUNCTION_PTR_HPP)
+#define BOOST_PHOENIX_PREPROCESSED_BIND_DETAIL_FUNCTION_PTR_HPP
 
 #if BOOST_PHOENIX_LIMIT <= 10
-#include <boost/phoenix/bind/preprocessed/bind_function_object_10.hpp>
+#include <boost/phoenix/bind/detail/cpp03/preprocessed/function_ptr_10.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 20
-#include <boost/phoenix/bind/preprocessed/bind_function_object_20.hpp>
+#include <boost/phoenix/bind/detail/cpp03/preprocessed/function_ptr_20.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 30
-#include <boost/phoenix/bind/preprocessed/bind_function_object_30.hpp>
+#include <boost/phoenix/bind/detail/cpp03/preprocessed/function_ptr_30.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 40
-#include <boost/phoenix/bind/preprocessed/bind_function_object_40.hpp>
+#include <boost/phoenix/bind/detail/cpp03/preprocessed/function_ptr_40.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 50
-#include <boost/phoenix/bind/preprocessed/bind_function_object_50.hpp>
+#include <boost/phoenix/bind/detail/cpp03/preprocessed/function_ptr_50.hpp>
 #else
 #error "BOOST_PHOENIX_LIMIT out of bounds for preprocessed headers"
 #endif
