+++ include/boost/phoenix/core/limits.hpp
@@ -47,6 +47,8 @@
 # define BOOST_PHOENIX_NO_VARIADIC_FUNCTION_EVAL
 #endif
 
+# define BOOST_PHOENIX_NO_VARIADIC_BIND
+
 #if !defined(BOOST_PHOENIX_ARG_LIMIT)
 # define BOOST_PHOENIX_ARG_LIMIT BOOST_PHOENIX_PP_ROUND_UP(BOOST_PHOENIX_LIMIT)
 #elif (BOOST_PHOENIX_ARG_LIMIT < 5)
